requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
fonttools==4.55.6
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
multidict==6.1.0
mysql-connector-python==9.1.0
numpy==2.2.2
optional==0.0.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tzdata==2025.1
urllib3==2.3.0
XlsxWriter==3.2.1
yarl==1.18.3
